==> Collect_Chart_Lab_Events/01_diagnosis.ipynb <==
## Return diagnosis icd 9 & 10 info.

Reading icd codes from csv file

Import file -> CSV\Imports\icd_thrombo.csv

Export file -> CSV\Exports\stroke_diagnosis.csv

--------------------------------------------
In this part of the code we create
a file with the admissions that have
been diagnosed with the disease that
we have declared through the ICD codes
in the CSV file. At the end of the code,
the file is produced with the unique admissions
and not the movements within the hospital units.
If I want to search for another disease
I have to change the path of the icd variable
which you find five lines below.

import gc
import gzip
import pandas as pd

# CAUTION: Path with the specific ICD codes from a CSV file 
icd = 'CSV\Imports\icd_stroke.csv'

compressed_file_path = r"..\Datasets\mimic-iv-2_2\hosp\diagnoses_icd.csv.gz"

try:    
    # Read and pass the compressed CSV file into a DataFrame
    df = pd.read_csv(compressed_file_path, compression='gzip')
    
    # Read the specific ICD codes from a CSV file
    specific_icd_numbers = pd.read_csv(icd, header=None)
    
    # Convert the list of ICD codes to a list of strings
    specific_icd_numbers = specific_icd_numbers[0].tolist()
    
    # Use boolean indexing to filter the DataFrame
    filtered_df = df[df['icd_code'].isin(specific_icd_numbers)]
    filtered_df = filtered_df.drop_duplicates(subset=['hadm_id'])
    
except FileNotFoundError:
    print("File not found.")
except Exception as e:
    print("An error occurred:", e)

# Export the filtered DataFrame to a CSV file
filtered_df.to_csv('CSV\Exports\o01_diagnosis.csv', index=False)

# Free RAM
gc.collect()
