==> Collect_Chart_Lab_Events/04_ICU_merged_groupby.ipynb <==
## Chartevent Groupby 1st Phase

At this phase we create in seperate files the observations

Attention -> set the observation time window in the second cell

Create rows for the final table.

Calculate: Mean, Median, Min, Max

import pandas as pd
import os
import gc

# Read csv
df_patients_chart = pd.read_csv('CSV\Exports\o04_icu_chart_lab_merged.csv', low_memory=False)
row_df = pd.read_csv('CSV\Exports\o03_icu_first_stay.csv')

# Convert charttime column from object to datetime
df_patients_chart['charttime']= pd.to_datetime(df_patients_chart['charttime'])

"""
In this field we are setting the
observation time window. In the
variable "start_icrease_time" we set
the observation step. In variable
"end_time" we declare when we want to
stop the observations. We set those
variables to hours.
"""

# Initialize time window (start, increase and end)
start_increase_time = 3
end_time = 48

# Filter the DataFrame to keep only the desired columns
temp_df = row_df[['subject_id', 'hadm_id']]

# Repeat each row three times
temp_df = temp_df.loc[temp_df.index.repeat(end_time/start_increase_time)].reset_index(drop=True)

# Create the "Time_Zone" column and count the occurrences of subject_id
temp_df['Time_Zone'] = temp_df.groupby('subject_id').cumcount() + 1

# Save the resulting DataFrame to a new CSV file
temp_df.to_csv('CSV\Exports\o07_chartevent_rows.csv', index=False)

## Calculating Mean

# Initialize the starting time window
time_window = pd.Timedelta(hours=start_increase_time)

# Define the desired time window
desired_time_window = pd.Timedelta(hours=end_time)

# Initialize a counter to track iterations
iteration = 1

# Define the directory path for saving files
output_directory = 'CSV/Exports/groupby/chartevent/mean/'

# Create the directory if it doesn't exist
os.makedirs(output_directory, exist_ok=True)

# Loop through the time windows
while time_window <= desired_time_window:
    grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                                   'hadm_id',
                                                   'itemid',
                                                   pd.Grouper(key='charttime',
                                                              freq=time_window)])
                        .agg({'valuenum': ['mean']}))
    grouped_multiple.columns = ['Mean_Chart']
    grouped_multiple = grouped_multiple.reset_index()
    
    # Add a new column that counts from 1 for each unique combination
    grouped_multiple['Count'] = grouped_multiple.groupby(['subject_id', 'hadm_id', 'itemid']).cumcount() + 1
    
    # Add a 'File_Number' column with the current iteration value
    grouped_multiple['Time_Zone'] = iteration
    
    # Define the file path with forward slashes
    result_file_name = os.path.join(output_directory, f'results_iteration_mean_o{iteration}.csv')
    grouped_multiple.to_csv(result_file_name, index=False)
    
    # Increase the time window by 3 hours
    time_window += pd.Timedelta(hours=start_increase_time)
    
    # Increment the iteration counter
    iteration += 1

## Calculating Median

# Initialize the starting time window
time_window = pd.Timedelta(hours=start_increase_time)

# Define the desired time window
desired_time_window = pd.Timedelta(hours=end_time)

# Initialize a counter to track iterations
iteration = 1

# Define the directory path for saving files
output_directory = 'CSV/Exports/groupby/chartevent/median/'

# Create the directory if it doesn't exist
os.makedirs(output_directory, exist_ok=True)

# Loop through the time windows
while time_window <= desired_time_window:
    grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                                   'hadm_id',
                                                   'itemid',
                                                   pd.Grouper(key='charttime',
                                                              freq=time_window)])
                        .agg({'valuenum': ['median']}))
    grouped_multiple.columns = ['Median_Chart']
    grouped_multiple = grouped_multiple.reset_index()
    
    # Add a new column that counts from 1 for each unique combination
    grouped_multiple['Count'] = grouped_multiple.groupby(['subject_id', 'hadm_id', 'itemid']).cumcount() + 1

    # Add a 'File_Number' column with the current iteration value
    grouped_multiple['Time_Zone'] = iteration
    
    # Define the file path with forward slashes
    result_file_name = os.path.join(output_directory, f'results_iteration_median_o{iteration}.csv')
    grouped_multiple.to_csv(result_file_name, index=False)
    
    # Increase the time window by 3 hours
    time_window += pd.Timedelta(hours=start_increase_time)
    
    # Increment the iteration counter
    iteration += 1

## Calculating Min

# Initialize the starting time window
time_window = pd.Timedelta(hours=start_increase_time)

# Define the desired time window
desired_time_window = pd.Timedelta(hours=end_time)

# Initialize a counter to track iterations
iteration = 1

# Define the directory path for saving files
output_directory = 'CSV/Exports/groupby/chartevent/min/'

# Create the directory if it doesn't exist
os.makedirs(output_directory, exist_ok=True)

# Loop through the time windows
while time_window <= desired_time_window:
    grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                                   'hadm_id',
                                                   'itemid',
                                                   pd.Grouper(key='charttime',
                                                              freq=time_window)])
                        .agg({'valuenum': ['min']}))
    grouped_multiple.columns = ['Min_Chart']
    grouped_multiple = grouped_multiple.reset_index()
    
    # Add a new column that counts from 1 for each unique combination
    grouped_multiple['Count'] = grouped_multiple.groupby(['subject_id', 'hadm_id', 'itemid']).cumcount() + 1
    
    # Add a 'File_Number' column with the current iteration value
    grouped_multiple['Time_Zone'] = iteration
    
    # Define the file path with forward slashes
    result_file_name = os.path.join(output_directory, f'results_iteration_min_o{iteration}.csv')
    grouped_multiple.to_csv(result_file_name, index=False)
    
    # Increase the time window by 3 hours
    time_window += pd.Timedelta(hours=start_increase_time)
    
    # Increment the iteration counter
    iteration += 1

## Calculating Max

# Initialize the starting time window
time_window = pd.Timedelta(hours=start_increase_time)

# Define the desired time window
desired_time_window = pd.Timedelta(hours=end_time)

# Initialize a counter to track iterations
iteration = 1

# Define the directory path for saving files
output_directory = 'CSV/Exports/groupby/chartevent/max/'

# Create the directory if it doesn't exist
os.makedirs(output_directory, exist_ok=True)

# Loop through the time windows
while time_window <= desired_time_window:
    grouped_multiple = (df_patients_chart.groupby(['subject_id',
                                                   'hadm_id',
                                                   'itemid',
                                                   pd.Grouper(key='charttime',
                                                              freq=time_window)])
                        .agg({'valuenum': ['max']}))
    grouped_multiple.columns = ['Max_Chart']
    grouped_multiple = grouped_multiple.reset_index()
    
    # Add a new column that counts from 1 for each unique combination
    grouped_multiple['Count'] = grouped_multiple.groupby(['subject_id', 'hadm_id', 'itemid']).cumcount() + 1
    
    # Add a 'File_Number' column with the current iteration value
    grouped_multiple['Time_Zone'] = iteration
    
    # Define the file path with forward slashes
    result_file_name = os.path.join(output_directory, f'results_iteration_max_o{iteration}.csv')
    grouped_multiple.to_csv(result_file_name, index=False)
    
    # Increase the time window by 3 hours
    time_window += pd.Timedelta(hours=start_increase_time)
    
    # Increment the iteration counter
    iteration += 1

# Free RAM
gc.collect()
